refactoring path aliases pt.1
