+++ package.json
@@ -52,6 +52,7 @@
         "tailwindcss": "^3.3.3",
         "typescript": "^5.0.2",
         "vite": "^4.4.5",
+        "vite-tsconfig-paths": "^4.3.2",
         "vitest": "^0.34.6"
     }
 }
