+++ src/components/Button/Button.type.ts
@@ -1,0 +1,3 @@
+import type { ButtonStylesProps } from "./Button.styles";
+export type ButtonProps = ButtonStylesProps &
+    React.ButtonHTMLAttributes<HTMLButtonElement>;
