+++ src/components/PageWrapper/PageWrapper.tsx
@@ -1,7 +1,7 @@
 import { FC } from "react";
-import type { ComponentProps } from "@components";
+import type { ComponentProps } from "@components/types";
 import { useLocation } from "react-router-dom";
-import { titles } from "@utils";
+import { titles } from "@/navigation/constants";
 import { Navbar } from "@components";
 
 export const PageWrapper: FC<ComponentProps> = ({ children }) => {
