+++ src/components/types.ts
@@ -1,5 +1,13 @@
 export interface ComponentProps {
     children?: React.ReactNode;
 }
-
 export type { TextInputProps } from "./TextInput/TextInput";
+export type { ButtonStylesProps } from "./Button/Button.styles";
+export type { ButtonProps } from "./Button/Button.type";
+export type {
+    NotificationProps,
+    NotificationData,
+} from "./Notification/Notification";
+export type { ErrorAlertProps } from "./ErrorAlert/ErrorAlert";
+export type { SelectProps } from "./Select/Select";
+export type { MultiSelectProps } from "./MultiSelect/MultiSelect";
