+++ src/navigation/constants.ts
@@ -1,0 +1,1 @@
+export { routes, titles } from "./Router";
