+++ src/providers/types.ts
@@ -1,0 +1,1 @@
+export type { ProviderName } from "./auth.provider";
