--- src/utils/index.tsx
@@ -1,14 +1,0 @@
-/*
- *
- * @utility Router
- *
- */
-export { Router, routes, titles } from "./Router";
-export { ProtectedRoutes } from "./ProtectedRoutes/ProtectedRoutes";
-
-/*
- *
- * @utility Test Utils
- *
- */
-export { renderWithRouter } from "./test-utils";
